feat: Add link function to generate bytecode with references
This commit adds a new file, `_link.ts`, which contains the implementation of the `link` function. The `link` function takes in a bytecode string, an array of references, and either an array of addresses or an object of addresses. It replaces the placeholders in the bytecode string with the corresponding addresses and returns the modified bytecode.

The purpose of this change is to provide a convenient way to generate bytecode with references for EVM contracts. This will be useful for contract deployment and interaction.